Externalize version 2 state upgrader
