+++ azurecaf/resource_name.go
@@ -1,7 +1,6 @@
 package azurecaf
 
 import (
-	"context"
 	"fmt"
 	"regexp"
 	"strings"
@@ -9,103 +8,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
-
-func resourceNameV2() *schema.Resource {
-	resourceMapsKeys := make([]string, 0, len(ResourceDefinitions))
-	for k := range ResourceDefinitions {
-		resourceMapsKeys = append(resourceMapsKeys, k)
-	}
-	return &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"name": {
-				Type:     schema.TypeString,
-				Optional: true,
-				ForceNew: true,
-				Default:  "",
-			},
-			"prefixes": {
-				Type: schema.TypeList,
-				Elem: &schema.Schema{
-					Type:         schema.TypeString,
-					ValidateFunc: validation.NoZeroValues,
-				},
-				Optional: true,
-				ForceNew: true,
-			},
-			"suffixes": {
-				Type: schema.TypeList,
-				Elem: &schema.Schema{
-					Type:         schema.TypeString,
-					ValidateFunc: validation.NoZeroValues,
-				},
-				Optional: true,
-				ForceNew: true,
-			},
-			"random_length": {
-				Type:         schema.TypeInt,
-				Optional:     true,
-				ForceNew:     true,
-				ValidateFunc: validation.IntAtLeast(0),
-				Default:      0,
-			},
-			"result": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"results": {
-				Type: schema.TypeMap,
-				Elem: &schema.Schema{
-					Type: schema.TypeString,
-				},
-				Computed: true,
-			},
-			"separator": {
-				Type:     schema.TypeString,
-				Optional: true,
-				ForceNew: true,
-				Default:  "-",
-			},
-			"clean_input": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				ForceNew: true,
-				Default:  true,
-			},
-			"passthrough": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				ForceNew: true,
-				Default:  false,
-			},
-			"resource_type": {
-				Type:         schema.TypeString,
-				Optional:     true,
-				ValidateFunc: validation.StringInSlice(resourceMapsKeys, false),
-				ForceNew:     true,
-			},
-			"resource_types": {
-				Type: schema.TypeList,
-				Elem: &schema.Schema{
-					Type:         schema.TypeString,
-					ValidateFunc: validation.StringInSlice(resourceMapsKeys, false),
-				},
-				Optional: true,
-				ForceNew: true,
-			},
-			"random_seed": {
-				Type:     schema.TypeInt,
-				Optional: true,
-				ForceNew: true,
-			},
-		},
-	}
-}
-
-func resourceNameStateUpgradeV2(ctx context.Context, rawState map[string]interface{}, meta interface{}) (map[string]interface{}, error) {
-	rawState["use_slug"] = true
-
-	return rawState, nil
-}
 
 func resourceName() *schema.Resource {
 	resourceMapsKeys := make([]string, 0, len(ResourceDefinitions))
